refactor  updatesetting view by using mvvm and databinding and livedata
